+++ src/main/java/cd/go/contrib/elasticagents/docker/DockerContainer.java
@@ -96,7 +96,7 @@ public class DockerContainer {
         }
 
         ContainerConfig.Builder containerConfigBuilder = ContainerConfig.builder();
-        if (request.properties().containsKey("Command")) {
+        if (StringUtils.isNotBlank(request.properties().get("Command"))) {
             containerConfigBuilder.cmd(splitIntoLinesAndTrimSpaces(request.properties().get("Command")).toArray(new String[]{}));
         }
 
@@ -120,7 +120,9 @@ public class DockerContainer {
         Set<String> env = new HashSet<>();
 
         env.addAll(settings.getEnvironmentVariables());
-        env.addAll(splitIntoLinesAndTrimSpaces(request.properties().get("Environment")));
+        if (StringUtils.isNotBlank(request.properties().get("Environment"))) {
+            env.addAll(splitIntoLinesAndTrimSpaces(request.properties().get("Environment")));
+        }
 
         env.addAll(Arrays.asList(
                 "GO_EA_MODE=" + mode(),