+++ src/main/java/cd/go/contrib/elasticagents/docker/DockerPlugin.java
@@ -16,10 +16,9 @@
 
 package cd.go.contrib.elasticagents.docker;
 
-import cd.go.contrib.elasticagents.docker.executors.GetPluginConfigurationExecutor;
-import cd.go.contrib.elasticagents.docker.executors.GetViewRequestExecutor;
-import cd.go.contrib.elasticagents.docker.executors.ServerPingRequestExecutor;
+import cd.go.contrib.elasticagents.docker.executors.*;
 import cd.go.contrib.elasticagents.docker.requests.CreateAgentRequest;
+import cd.go.contrib.elasticagents.docker.requests.ProfileValidateRequest;
 import cd.go.contrib.elasticagents.docker.requests.ShouldAssignWorkRequest;
 import cd.go.contrib.elasticagents.docker.requests.ValidatePluginSettings;
 import com.thoughtworks.go.plugin.api.GoApplicationAccessor;
@@ -62,6 +61,14 @@ public class DockerPlugin implements GoPlugin {
                     return new ServerPingRequestExecutor(agentInstances, pluginRequest).execute();
                 case PLUGIN_SETTINGS_GET_VIEW:
                     return new GetViewRequestExecutor().execute();
+                case REQUEST_GET_PROFILE_METADATA:
+                    return new GetProfileMetadataExecutor().execute();
+                case REQUEST_GET_PROFILE_VIEW:
+                    return new GetProfileViewExecutor().execute();
+                case REQUEST_VALIDATE_PROFILE:
+                    return ProfileValidateRequest.fromJSON(request.requestBody()).executor().execute();
+                case PLUGIN_SETTINGS_GET_ICON:
+                    return new GetPluginSettingsIconExecutor().execute();
                 case PLUGIN_SETTINGS_GET_CONFIGURATION:
                     return new GetPluginConfigurationExecutor().execute();
                 case PLUGIN_SETTINGS_VALIDATE_CONFIGURATION: