+++ src/main/java/cd/go/contrib/elasticagents/docker/utils/Util.java
@@ -19,9 +19,11 @@ package cd.go.contrib.elasticagents.docker.utils;
 import cd.go.contrib.elasticagents.docker.executors.GetViewRequestExecutor;
 import com.google.common.base.Function;
 import com.google.common.collect.Collections2;
+import com.google.common.io.ByteStreams;
 import com.google.common.io.CharStreams;
 
 import java.io.IOException;
+import java.io.InputStream;
 import java.io.InputStreamReader;
 import java.io.StringReader;
 import java.nio.charset.StandardCharsets;
@@ -37,6 +39,14 @@ public class Util {
     public static String readResource(String resourceFile) {
         try (InputStreamReader reader = new InputStreamReader(GetViewRequestExecutor.class.getResourceAsStream(resourceFile), StandardCharsets.UTF_8)) {
             return CharStreams.toString(reader);
+        } catch (IOException e) {
+            throw new RuntimeException("Could not find resource " + resourceFile, e);
+        }
+    }
+
+    public static byte[] readResourceBytes(String resourceFile) {
+        try (InputStream in = GetViewRequestExecutor.class.getResourceAsStream(resourceFile)) {
+            return ByteStreams.toByteArray(in);
         } catch (IOException e) {
             throw new RuntimeException("Could not find resource " + resourceFile, e);
         }