+++ src/test/java/cd/go/contrib/elasticagents/docker/executors/GetPluginSettingsIconExecutorTest.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright 2016 ThoughtWorks, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package cd.go.contrib.elasticagents.docker.executors;
+
+import cd.go.contrib.elasticagents.docker.utils.Util;
+import com.google.gson.Gson;
+import com.thoughtworks.go.plugin.api.response.GoPluginApiResponse;
+import org.glassfish.jersey.internal.util.Base64;
+import org.junit.Test;
+
+import java.util.HashMap;
+
+import static org.hamcrest.CoreMatchers.is;
+import static org.junit.Assert.assertThat;
+
+public class GetPluginSettingsIconExecutorTest {
+
+    @Test
+    public void rendersIconInBase64() throws Exception {
+        GoPluginApiResponse response = new GetPluginSettingsIconExecutor().execute();
+        HashMap<String, String> hashMap = new Gson().fromJson(response.responseBody(), HashMap.class);
+        assertThat(hashMap.size(), is(2));
+        assertThat(hashMap.get("content-type"), is("image/svg+xml"));
+        assertThat(Util.readResource("/docker-plain.svg"), is(Base64.decodeAsString(hashMap.get("data"))));
+    }
+}